updated Main console class . RUN the main console

+++ src/Account.java
@@ -41,7 +41,10 @@ public class Account {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
+
+
+
 }
